doc: module documentation
